Move core interfaces into contracts
