+++ src/Monofony/Bridge/Sylius/User/EventListener/PasswordUpdaterListener.php
@@ -11,7 +11,7 @@
 
 namespace Monofony\Bridge\Sylius\User\EventListener;
 
-use Monofony\Component\Core\Model\Customer\CustomerInterface;
+use Monofony\Contracts\Core\Model\Customer\CustomerInterface;
 use Sylius\Bundle\UserBundle\EventListener\PasswordUpdaterListener as BasePasswordUpdaterListener;
 use Sylius\Component\Resource\Exception\UnexpectedTypeException;
 use Symfony\Component\EventDispatcher\GenericEvent;
