+++ src/Monofony/Contracts/Core/LICENSE
