+++ src/Monofony/Pack/AdminPack/Recipe/tests/Behat/Context/Ui/Backend/ManagingAdministratorsContext.php
@@ -11,7 +11,7 @@ use App\Tests\Behat\Page\Backend\Administrator\IndexPage;
 use App\Tests\Behat\Page\Backend\Administrator\UpdatePage;
 use Behat\Behat\Context\Context;
 use Doctrine\Common\Persistence\ObjectManager;
-use Monofony\Component\Core\Model\User\AdminUserInterface;
+use Monofony\Contracts\Core\Model\User\AdminUserInterface;
 use Monofony\Bridge\Behat\Service\NotificationCheckerInterface;
 use Monofony\Bridge\Behat\Service\SharedStorageInterface;
 use Webmozart\Assert\Assert;
