+++ src/Monofony/Pack/ApiPack/Recipe/src/Provider/CustomerProvider.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Provider;
 
-use Monofony\Component\Core\Model\Customer\CustomerInterface;
+use Monofony\Contracts\Core\Model\Customer\CustomerInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 use Sylius\Component\User\Canonicalizer\CanonicalizerInterface;
