+++ src/Monofony/Pack/ApiPack/Recipe/src/Provider/CustomerProviderInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Provider;
 
-use Monofony\Component\Core\Model\Customer\CustomerInterface;
+use Monofony\Contracts\Core\Model\Customer\CustomerInterface;
 
 interface CustomerProviderInterface
 {
