+++ src/Monofony/Pack/CorePack/Recipe/spec/App/EventSubscriber/UserRegistrationSubscriberSpec.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace spec\App\EventSubscriber;
 
-use Monofony\Component\Core\Model\Customer\CustomerInterface;
+use Monofony\Contracts\Core\Model\Customer\CustomerInterface;
 use Doctrine\Common\EventSubscriber;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\ORM\Events;
