+++ src/Monofony/Pack/CorePack/Recipe/spec/App/Form/EventSubscriber/CustomerRegistrationFormSubscriberSpec.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace spec\App\Form\EventSubscriber;
 
-use Monofony\Component\Core\Model\Customer\CustomerInterface;
-use Monofony\Component\Core\Model\User\AppUserInterface;
+use Monofony\Contracts\Core\Model\Customer\CustomerInterface;
+use Monofony\Contracts\Core\Model\User\AppUserInterface;
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
