+++ src/Monofony/Pack/CorePack/Recipe/src/Context/CustomerContext.php
@@ -2,7 +2,7 @@
 
 namespace App\Context;
 
-use Monofony\Component\Core\Model\User\AppUserInterface;
+use Monofony\Contracts\Core\Model\User\AppUserInterface;
 use Sylius\Component\Customer\Context\CustomerContextInterface;
 use Sylius\Component\Customer\Model\CustomerInterface;
 use Sylius\Component\User\Model\UserInterface;
