+++ src/Monofony/Pack/CorePack/Recipe/src/Entity/Customer/Customer.php
@@ -3,8 +3,8 @@
 namespace App\Entity\Customer;
 
 use Doctrine\ORM\Mapping as ORM;
-use Monofony\Component\Core\Model\Customer\CustomerInterface;
-use Monofony\Component\Core\Model\User\AppUserInterface;
+use Monofony\Contracts\Core\Model\Customer\CustomerInterface;
+use Monofony\Contracts\Core\Model\User\AppUserInterface;
 use Sylius\Component\Customer\Model\Customer as BaseCustomer;
 use Sylius\Component\User\Model\UserInterface;
 use Symfony\Component\Validator\Constraints as Assert;
