+++ src/Monofony/Pack/CorePack/Recipe/src/Entity/User/AdminAvatar.php
@@ -6,7 +6,7 @@ namespace App\Entity\User;
 
 use App\Entity\Media\File;
 use Doctrine\ORM\Mapping as ORM;
-use Monofony\Component\Core\Model\User\AdminAvatarInterface;
+use Monofony\Contracts\Core\Model\User\AdminAvatarInterface;
 use Symfony\Component\Validator\Constraints as Assert;
 use Vich\UploaderBundle\Mapping\Annotation as Vich;
