+++ src/Monofony/Pack/CorePack/Recipe/tests/Behat/Context/Setup/AppSecurityContext.php
@@ -4,7 +4,7 @@ namespace App\Tests\Behat\Context\Setup;
 
 use Monofony\Bridge\Behat\Service\AppSecurityServiceInterface;
 use Monofony\Bridge\Behat\Service\SharedStorageInterface;
-use Monofony\Component\Core\Model\User\AppUserInterface;
+use Monofony\Contracts\Core\Model\User\AppUserInterface;
 use App\Fixture\Factory\AdminUserExampleFactory;
 use Behat\Behat\Context\Context;
 use Sylius\Component\User\Repository\UserRepositoryInterface;
