+++ src/Monofony/Pack/CorePack/Recipe/tests/Behat/Context/Setup/CustomerContext.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Tests\Behat\Context\Setup;
 
 use Monofony\Bridge\Behat\Service\SharedStorageInterface;
-use Monofony\Component\Core\Model\Customer\CustomerInterface;
-use Monofony\Component\Core\Model\User\AppUserInterface;
+use Monofony\Contracts\Core\Model\Customer\CustomerInterface;
+use Monofony\Contracts\Core\Model\User\AppUserInterface;
 use Behat\Behat\Context\Context;
 use Doctrine\Common\Persistence\ObjectManager;
 use Sylius\Component\Resource\Factory\FactoryInterface;
