+++ src/Monofony/Pack/CorePack/Recipe/tests/Behat/Context/Setup/UserContext.php
@@ -6,7 +6,7 @@ use App\Fixture\Factory\AppUserExampleFactory;
 use Behat\Behat\Context\Context;
 use Doctrine\Common\Persistence\ObjectManager;
 use Monofony\Bridge\Behat\Service\SharedStorageInterface;
-use Monofony\Component\Core\Model\User\AppUserInterface;
+use Monofony\Contracts\Core\Model\User\AppUserInterface;
 use Sylius\Component\User\Model\UserInterface;
 use Sylius\Component\User\Repository\UserRepositoryInterface;
